image_processor/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pika==1.3.2
Pillow==10.0.0
psycopg2-binary==2.9.6
requests==2.31.0
SQLAlchemy==2.0.18
SQLAlchemy-Utils==0.41.1
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
